V1.0.9

+++ hostel/authentication/users/views.py
@@ -8,9 +8,11 @@ from .serializers import CustomUserSerializer
 
 
 class CustomUserView(APIView):
-    def get(self, request, pk):
+    def get(self, request):
+        if not request.user.id:
+            return Response('unauthorized', 401)
         print(request.headers)
-        user = get_object_or_404(CustomUser.objects.all(), pk=pk)
+        user = get_object_or_404(CustomUser.objects.all(), pk=request.user.id)
         serializer = CustomUserSerializer(user)
         user_data = serializer.data
         user_data.pop('password', None)
@@ -27,18 +29,23 @@ class CustomUserView(APIView):
             "user": serializer.data
         })
 
-    def put(self, request, pk):
-        saved_user = get_object_or_404(CustomUser.objects.all(), pk=pk)
+    def put(self, request):
+        if not request.user.id:
+            return Response('unauthorized', 401)
+        saved_user = get_object_or_404(CustomUser.objects.all(), pk=request.user.id)
         serializer = CustomUserSerializer(instance=saved_user, data=request.data, partial=True)
         if serializer.is_valid(raise_exception=True):
             user_saved = serializer.save()
-        return Response({
-            "user": serializer.data
-        })
+            return Response({
+                "user": serializer.data
+            })
+        return Response(f"Invalid data", 501)
 
-    def delete(self, request, pk):
-        user = get_object_or_404(CustomUser.objects.all(), pk=pk)
+    def delete(self, request):
+        if not request.user.id:
+            return Response('unauthorized', 401)
+        user = get_object_or_404(CustomUser.objects.all(), pk=request.user.id)
         user.delete()
         return Response({
-            "message": f"User with id '{pk}' has been deleted."
+            "message": f"User with id '{request.user.id}' has been deleted."
         }, status=204)
